requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
diffusers>=0.31.0
transformers>=4.49.0
accelerate>=1.1.1
ftfy
imageio
imageio-ffmpeg
numpy
pillow
opencv-python>=4.9.0.80
tokenizers>=0.20.3
tqdm
easydict
huggingface-hub
safetensors
xformers
flask>=2.3.0
werkzeug>=2.3.0 